order getPreparationtime renamed -> getProductiontime for consistency getDuration returns int instead of double now for consistency with other time calculations that use int renamed be.kdb.distanceAPI.Position -> Point for consistency with start code ordermanager.assignorder bug: no save to repo after o.setDeliverer removed package foundation/operatie

+++ src/be/kdg/deliDish/business/delivery/Move.java
@@ -6,5 +6,5 @@ package be.kdg.deliDish.business.delivery;
 public interface Move {
     double getDistance();
 
-    double getDuration();
+    int getDuration();
 }
